Add support for configuring plugins with values from settings
More importantly this achieves the above without having to make
breaking changes to the plugin api to support promises.

Ideally the configure, customise and use calls for the plugin api
would support Promise<T> instead of T through the whole process but
this is not currently the case and would be a breaking change.

This system allows configuration via settings / config while only
requiring that use return a promise, which is already supported.

This does require that plugins are coded to call resolveSettings
but this is a relatively easy change, and will be demonstrated in
ubiquity-client.